refactor: Update TabletSignInView with authentication functionality and form validation
This commit refactors the TabletSignInView widget to include authentication functionality and form validation. The following changes were made:

Added state management for the authentication process using the isAuthenticating state provider.
Created a GlobalKey for the form to enable form validation.
Initialized and disposed of TextEditingController instances for the user email address and password fields.
Updated the elevated button's onPressed callback to call the signIn function from the signIn_services.dart file, passing the user email and password.
Added form validation logic to the email and password fields using custom validators.
Implemented an AuthenticatingSpinnerWidget to display a spinner while authentication is in progress.
Updated the customDialogBox function call to display a dialog box with a message about the Google Sign-In feature not being implemented yet.
These changes enhance the functionality of the sign-in screen, allowing users to authenticate with their credentials and providing appropriate error messages and feedback during the authentication process.