Refactor: Enhanced customDialogBox function
The customDialogBox function has been updated to improve its flexibility and usability. The following changes have been made:

The width and height parameters are now nullable (double?). This change allows the function to be used with default dimensions if specific values are not provided.

The default width and height values have been set to 300 if no specific values are provided.

The isWithCloseIconButton parameter now has a default value of true. This means that by default, the dialog box will display a close icon button. However, this behavior can be overridden by passing false as the argument.

The widget parameter is now nullable (Widget?). This change allows for an optional additional widget to be displayed in the dialog box.

The closeButtonOnPressed parameter has been renamed to closeButtonPressed. This change aligns the parameter name with its purpose more clearly.

These enhancements provide more flexibility in configuring the custom dialog box and make the function more user-friendly.