feat: Integrate Flutter's animate package for enhanced animations
This commit introduces the integration of the Flutter animate package to the project. By incorporating this package, we can enhance the animations within our application, resulting in a more visually appealing and interactive user experience. This addition will provide a wide range of animation options and features that can be leveraged throughout the application.